+++ src/index.ts
@@ -1,6 +1,11 @@
 export { inject } from './inject';
 export { ProxyDiContainer } from './ProxyDiContainer';
 export { injectable } from './injectable';
+export { middleware } from './middleware/middleware';
+export {
+    MiddlewareRegisteringListener,
+    MiddlewareRemovingListener,
+} from './middleware/MiddlewareListener';
 export {
     Injection,
     DependencyId,