+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [[0.0.10](https://www.npmjs.com/package/proxydi/v/0.0.11)] - 2025-03-03
+## [[0.0.12](https://www.npmjs.com/package/proxydi/v/0.0.12)] - 2025-03-04
 
 ### Added
 
-- resolveInjectables() method [#12](https://github.com/proxy-di/proxydi/pull/13)
+- Experimental @middleware [#14](https://github.com/proxy-di/proxydi/pull/14)
+
+## [[0.0.11](https://www.npmjs.com/package/proxydi/v/0.0.11)] - 2025-03-03
+
+### Added
+
+- resolveInjectables() method [#13](https://github.com/proxy-di/proxydi/pull/13)
 
 ## [[0.0.10](https://www.npmjs.com/package/proxydi/v/0.0.10)] - 2025-03-02